create component for genre form using mui
